refactor(core): remove verified zero-reference legacy facade shims
