--- core/src/services/EmailService.ts
@@ -1,4 +1,0 @@
-/**
- * @deprecated Moved to bounded context
- */
-export * from '../domains/notifications/application/EmailService';
