--- core/src/services/RevenueAnalytics.ts
@@ -1,3 +1,0 @@
-// @deprecated
-// Remove during M6 or M8
-export * from "../domains/payments/application/RevenueAnalytics";
